refactor(class): change class name and logic
