+++ src/main/java/baseball/Application.java
@@ -1,10 +1,9 @@
 package baseball;
 
 public class Application {
-    
+
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        Game game = new Game();
-        game.startGame();
+        Game.startGame();
     }
 }
